restructured the bucketlist components
